Classification voting, fixing a bug, forwarding data

+++ modules/construction/cpt_reconstruction/include/cpt_reconstruction/reconstruction_model.h
@@ -79,7 +79,7 @@ class Model {
   std::vector<double>* getRadius();
   std::vector<int>* getShapeIDs();
   int getOutlierCount();
-  float getMinDistance();
+  double getMinDistance();
   void clearRansacShapes();
   void clearBuffer();
